Add more backfill and adjust time.
The coordinator may once in a while catch an edge case so adding an extra day for backfill may help. Home assistant also reports usage at the top of the hour, but forward looking while the data itself is backwards looking so adjusting it an hour back means the display will be between X-1 and X instead of X and X+1 which is the wrong time.